Move mentor invites to new folder and add vars
