+++ app/(api)/_actions/emails/sendSingleMentorInvite.ts
@@ -1,7 +1,7 @@
 'use server';
 
 import getOrCreateTitoInvitation from '@actions/tito/getOrCreateTitoInvitation';
-import mentorInviteTemplate from './emailFormats/2026MentorInviteTemplate';
+import mentorInviteTemplate from './emailTemplates/2026MentorInviteTemplate';
 import { DEFAULT_SENDER, transporter } from './transporter';
 import { MentorInviteData, SingleMentorInviteResponse } from '@typeDefs/emails';
